server/observability: refactor Slack utilities
